requirements.txt
langchain
langgraph
ipykernel
python-dotenv
langchain-google-genai
langchain-openai
uvicorn
fastapi
langchain-core
streamlit
langchain_community
langserve
sse_starlette
langchain_ollama
pydantic